testdata-generator-python/requirements.txt
# Test Data Generator Dependencies
faker>=22.0.0
click>=8.1.0
pyyaml>=6.0.0
